requirements.txt
pytest
requests
allure-pytest
playwright
pytest-playwright